Test that a remount can't discard unsaved YAML changes
The ConfigLoadTracker spec pinned the tracker's own API, but every assertion
in it passed against the ref-based implementation too -- so nothing actually
covered the data-loss bug.

Testing it needs the real InitialDataLoader, and main.tsx calls createRoot at
module scope, so importing it mounts the app. Extracted the component to
components/InitialDataLoader.tsx; main.tsx keeps only the entry point and the
App shell, dropping to 97 lines from 279.

The regression test mounts with the config query already resolved (the cached
state a remount sees), unmounts, remounts, and asserts the store is
initialised exactly once. Verified it earns its place: restoring the two refs
fails that test and leaves the other two green.
